support 64bit a1size

+++ HACApK_FPGA.c
@@ -13,7 +13,7 @@
  register int ip,il,it;
  int nlf,ndl,ndt,nstrtl,nstrtt,kt,itl,itt,ill;
  int st_lf_stride = st_leafmtxp->st_lf_stride;
- int a1size;
+ int64_t a1size;
  
  nlf=st_leafmtxp->nlf;
   //fprintf(stderr,"nlf=%d \n",nlf);
@@ -71,7 +71,7 @@
  register int ip,il,it;
  int nlf,ndl,ndt,nstrtl,nstrtt,kt,itl,itt,ill;
  int st_lf_stride = st_leafmtxp->st_lf_stride;
- int a1size;
+ int64_t a1size;
  int ith, nths, nthe;
  int ls, le;
  double *zaut, *zbut;
